+++ lib/src/modules/feedback_page/controllers/controllers.dart
@@ -1,0 +1,1 @@
+export 'feedback_controller.dart';
