+++ lib/src/modules/feedback_page/widgets/atoms/atoms.dart
@@ -1,0 +1,2 @@
+export 'feedback_or_report_statuses_card.dart';
+export 'feedback_type_button.dart';
