+++ lib/src/modules/feedback_page/widgets/organisms/organisms.dart
@@ -1,0 +1,1 @@
+export 'feedback_detail_form.dart';
