+++ lib/src/modules/feedback_page/widgets/pages/pages.dart
@@ -1,0 +1,1 @@
+export 'feedback_and_report_setting_page.dart';
